Implemented Util tests, and removed the function 'compare', to compare two doubles, from util; using the built in compareTo function instead now.

+++ src/main/java/util/TechnicalAnalysis.java
@@ -45,7 +45,7 @@ public class TechnicalAnalysis {
             i--;
         }
 
-        Double movingAverage = pricesSum / ((Double) numberOfDays);
+        Double movingAverage = pricesSum / ((Integer) numberOfDays).doubleValue();
 
         return movingAverage;
     }
